requirements.txt
streamlit
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
google-auth
python-dotenv
httpx
python-multipart
python-dateutil
requests
PyPDF2
python-docx
openpyxl